MRIGradients.GirfEssential — Method

GirfEssential(data::AbstractVecOrMat, vect::AbstractVector, isFreq::Bool, inChannels, outBasis)

Function to create a girf structure with the same interface as the constructor in the MATLAB implementation.

Arguments

data::AbstractVecOrMat - [nSamples x nOutBasis x nInChannels] Full GIRF data.

vect::AbstractVector - [nSamples] Frequency vector indicating the frequency range, in unit of Hz.

isFreq::Bool - True for GIRF in frequency domain, false for GIRF in temporal domain.

inChannels - Input gradient and shim channel names.

outBasis - Output field basis.

Outputs

g::GirfEssential - Constructed GirfEssential object.

MRIGradients.convertDomain! — Method

convertDomain!(g::GirfEssential)

Convert frequency-domain GIRF to time domain and vice versa. Also calculates dependent parameters.

Arguments

g::GirfEssential - The input GirfEssential object; conversion type is determined by g.isFreqDomainGirf

Outputs

g::GirfEssential - Modified GirfEssential object with converted/calculated fields.

This function was created from part of a code package for GIRF computation and application in MATLAB. The package is available under a BSD 3-clause license. Further info see: https://github.com/MRI-gradient/girf Original MATLAB Author: Johanna Vannesjo (johanna.vannesjo@gmail.com) Copyright (C) 2014 IBT, University of Zurich and ETH Zurich, 2016 FMRIB centre, University of Oxford

MRIGradients.displayGirf — Method

displayGirf(g::GirfEssential)

Function to print out the girf data in a human-readable format as a summary to make sure data is loaded properly

Arguments

g::GirfEssential - GirfEssential structure containing GIRF data

MRIGradients.loadGIRFMatlabTim — Function

GIRFfreq, GIRFdata, GIRF_length = loadGIRFTimMatlab() loads different measured GIRFs from Siemens Prisma, as computed in Matlab (Zhe "Tim" Wu's code)

Arguments

idGirf - identifier for selected GIRF 0 = example GIRF in this repo 1 = Nov 2020 pos blips 2 = June 2021 pos blips 3 = June 2021 pos/neg blips (4 averages) 4 = November 2021, pos/neg blips (4 averages)

Outputs

GIRF_freq -

GIRF_data - x 3 data

GIRF_length -

MRIGradients.loadGirf — Function

loadGirf(degree, id)

Function to load girf with specified id according to hardcoded file structure (internal use only)

Arguments

degree - specifies degree of the girf (1st or 0th order correction)

id - specifies girf measurement id since there have been multiple measurements

MRIGradients.readGIRFFile — Method

readGIRFFile(g::GirfEssential, pathX::String, pathY::String, pathZ::String, varName::String, doFilter::Bool)

Function to read GIRF files with specific variable name and return a corresponding GIRFEssential object.

Arguments

pathX::String - Full path for GIRF MAT-file on X axis

pathY::String - Full path for GIRF MAT-file on Y axis

pathZ::String - Full path for GIRF MAT-file on Z axis

varName::String - Variable name that needs to be read as GIRF data from the MAT-files.

doFilter::Bool - Whether we filter the GIRF data using Tukey filter

Outputs

g::GirfEssential - Constructed GirfEssential object with read-in data.

MRIGradients.setIdentifier! — Method

setIdentifier!(g::GirfEssential, identifier::String)

Function to name the GIRF with an identifier

Arguments

g::GirfEssential - GirfEssential object needs the identifier to be set.

identifier::String - Input identifier.

Outputs

g::GirfEssential - GirfEssential object with the identifier set as the input.