MRIGradients.gradients_to_nodes — Method

gradients_to_nodes(gradients::Matrix; gamma=42577478, dwellTime=2e-6, FOV=[220,220,1],reconSize=[200,200,1])

Function to calculate the k-space nodes from an applied gradient train

Arguments

gradients::Matrix - input gradients played out, acquired with constant dwell time. Size: (2 x N) where N is the number of gradient samples

gamma - gyromagnetic ratio [Hz]

dwellTime - the dwell time of the trajectory [s]

FOV - spatial field of view in [mm] [1 x 3]

reconSize - reconstruction matrix size [1 x 3]

Outputs

nodes::Matrix - k-space nodes corresponding to applied gradient train. Size: (2 x N) where N is the number of gradient samples

MRIGradients.nodes_to_gradients — Method

nodes_to_gradients(nodes::Matrix; gamma=42577478, dwellTime=2e-6, FOV=[220,220,1],reconSize=[200,200,1])

Function to calculate the gradient train used to acquire k-space data

Arguments

nodes::Matrix - input k-space nodes traversed in order, acquired with constant dwell time. Size: (2 x N) where N is the number of nodes

gamma - gyromagnetic ratio [Hz]

dwellTime - the dwell time of the trajectory [s]

FOV - spatial field of view in [mm] [1 x 3]

reconSize - reconstruction matrix size [1 x 3]

Outputs

gradients::Matrix - gradients corresponding to k-space samples. Size: (2 x N) where N is the number of k-space samples

MRIGradients.time2freq — Method

time2freq(t)

converts time domain vector into a vector of frequencies according to the Nyquist criterion Port of Johanna Vannesjo's time2freq method: https://github.com/MRI-gradient/GIRF/blob/main/utils/time2freq.m

Arguments

t - input time vector

Outputs

f - Calculated requency axis, vector with [length(t)]

df - Frequency resolution

f_max - Full frequency range. The frequency axis range is [-fmax/2, fmax/2].